Removal of inactive services from testnet docs.
There were a lot of services which are no longer active. These services have been removed.
A faucet, multiple explorers, and P2Pool nodes were removed.

In addition, line 64 was changed to adapt to the removal of the bamboo.dash.org service.

The latest binaries section was updated, removing the bamboo.dash.org services, replacing it with the wallets page on Dash.org

Content itself was not updated for 12.3, however this would serve as a better base for someone who wishes to do that.