build: pin Babel version (#236)
The latest version of Babel has an issue that causes an error when "make html" is run in some cases. Installing pytz seems to fix it, but this issue makes it seem like that dependency is undesirable: https://github.com/python-babel/babel/issues/716.

Pinning to Babel 2.11.0 for now avoids local build issues.